Applies CS fixes and stan etc
