+++ lib/Adapter/DTL/DTLArgumentResolver.php
@@ -2,8 +2,6 @@
 
 namespace Phpactor\LanguageServer\Adapter\DTL;
 
-use Amp\CancellationToken;
-use Amp\CancellationTokenSource;
 use DTL\ArgumentResolver\ArgumentResolver as UpstreamArgumentResolver;
 use DTL\ArgumentResolver\ParamConverter\RecursiveInstantiator;
 use Phpactor\LanguageServer\Core\Dispatcher\ArgumentResolver;
