--- lib/Core/Server/Parser/StreamParser.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Phpactor\LanguageServer\Core\Server\Parser;
-
-use Amp\Promise;
-
-interface StreamParser
-{
-    public function wait(): Promise;
-}
